# supplementary_files/swagger documentation.yaml
swagger: '2.0'
basePath: /
paths:
  /api/add_expense:
    post:
      responses:
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '402':
          description: Missing transaction details
          schema:
            $ref: '#/definitions/Error'
        '400':
          description: Missing transaction type
          schema:
            $ref: '#/definitions/Error'
        '201':
          description: Transaction created successfully
      description: Add a transaction.
      operationId: add_expense
      security:
      - BearerAuth: []
      tags:
      - transaction
...
  /api/create_passcode:
    post:
      responses:
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: No passcode entered
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '201':
          description: Passcode created successfully
      description: Create the passcode.
      operationId: Create Passcode
      security:
      - BearerAuth: []
      tags:
      - Expense Tracker Passcode
...
  /api/get_all_categories:
    get:
      responses:
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '202':
          description: Success
      description: Retrieve all transaction categories.
      operationId: get_all_categories
      security:
      - BearerAuth: []
      tags:
      - transaction
  /api/get_all_expense:
    get:
      responses:
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '200':
          description: Success
      description: Retrieve all transactions.
      operationId: get_all_transactions
      security:
      - BearerAuth: []
      tags:
      - transaction
...
  /api/get_passcode_status:
    get:
      responses:
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '202':
          description: Status retrieved
      description: Check if a passcode exists.
      operationId: Check Passcode Status
      security:
      - BearerAuth: []
      tags:
      - Expense Tracker Passcode
...
  /api/submit_passcode:
    post:
      responses:
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: No passcode entered
          schema:
            $ref: '#/definitions/Error'
        '405':
          description: No passcode found
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        '200':
          description: Success
      description: Submit the passcode.
      operationId: Submit Passcode
      security:
      - BearerAuth: []
      tags:
      - Expense Tracker Passcode
...
info:
  title: GrowUp Richie Swagger API Documentation
  version: '1.0'
  description: API for a financial literacy app for children aged 8-14
produces:
- application/json
consumes:
- application/json
securityDefinitions:
  BearerAuth:
    type: apiKey
    in: header
    name: Authorization
    description: Enter your JWT token with the `Bearer ` prefix, e.g., "Bearer <your_token>"
security:
- BearerAuth: []
...
definitions:
...
  Error:
    properties:
      error:
        type: string
        description: Error message
    type: object
...